orderhistoryandonlinePayment

+++ app/src/main/java/com/comunisolve/newmultiplerestaurantsapp/Database/CartDataSource.java
@@ -18,7 +18,7 @@ public interface CartDataSource {
 
     Single<Integer> countItemInCart(String userPhone, int restaurantId);
 
-    Single<Integer> sumPrice(String userPhone, int restaurantId);
+    Single<Long> sumPrice(String userPhone, int restaurantId);
 
     Single<CartItem> getItemInCart(String foodId, String userPhone, int restaurantId);
